finished miceblocks.c, although it does not solve yet.  Added it to the Makefile and I am not sure why m1ton is being commited.  I did not make any changes to it. - Damian Hites
git-svn-id: https://gamescrafters.svn.sourceforge.net/svnroot/gamescrafters/GamesmanClassic/trunk@971 a668be5d-ef11-cb4c-8bfd-e08cbb24f2ba

+++ gamesman/src/m1ton.c
@@ -386,7 +386,7 @@ USERINPUT GetAndPrintPlayersMove(thePosition, theMove, playerName)
 BOOLEAN ValidTextInput(input)
      STRING input;
 {
-  unsigned int theInput = atof(input);
+  unsigned int theInput = atoi(input);
   return(theInput <= gTurn && theInput >= 1);
 }
